Add function to change the http outbound url
Summary:
Add function to change the http outbound url

http.NewOutbound() returns a interface, and the underlying impl struct (which contains the URL field) is not exported by yarpc package

Thus add a helper function to set the URL using reflection.

This is needed as each peloton master has a http outbound pointing to the current leader. If the leader changes, the outbound URL need

to re-point to the new master url.

Test Plan: Testcase  added

Reviewers: min, avyas, igor, wu

Reviewed By: wu

Differential Revision: https://code.uberinternal.com/D617423