Add initial skeleton of hostmgr API and an empty server.
Summary:
This is aimed to get build system able to build a hostmgr instance instead of providing a functional complete component.

More API will be gradually added to hostmgr, and each API will be implemented in smaller diffs.

Test Plan: 1. make is able to generate new `bin/peloton-hostmgr` binary, which is executable and able to listen at a fixed 8084 port.

Reviewers: wu, min, xiaojian, #peloton_meta_framework, slu

Reviewed By: #peloton_meta_framework, slu

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D714446