Batch update/inserts for mysql and stapi
Summary:
Adds support for leveraging batched updates and inserts for both
the mysql and stapi storage systems. In mysql, multiple task creates are handled in a single
insert statement. In STAPI, there are n individual insert/update statements, and are all executed
in a single RPC. I think the stapi API is deficient here, in that it doesn't propogate individual
errors back for each api.Statement, instead, it only presents a single error.

NOTE: in my testing on my laptop, the batch create/update for tasks on mysql only took a few seconds for 10000 tasks.
@wu it would be great to put this code through its paces for real and get good comparison numbers against the previous version
and quantify how much this helps task launch speed.

NOTE: in attempting to test the stapi stuff, i realized the master does not ever seem to make a
connection to cassandra; it will always use the mysql connection. This seems wrong to me though. I must
be missing something

T613109

Reviewers: min, wu, igor, #peloton_meta_framework, slu

Reviewed By: igor, #peloton_meta_framework, slu

Subscribers: jenkins

Maniphest Tasks: T613109

Differential Revision: https://code.uberinternal.com/D707345