Make mesos task id unique by adding uuid part of it
Summary:
Make the mesos task id unique by adding a uuid as part of it

This is to make sure that in case a task is failed or lost, we can

launch the same peloton task with a different mesos task id. Otherwise

mesos would reject the re-launch as the task id is already used.

Test Plan: Existing test passes, and manually scheduled a job using local peloton cluster

Reviewers: min, wu, #peloton_meta_framework

Reviewed By: wu, #peloton_meta_framework

Differential Revision: https://code.uberinternal.com/D720673