Move offerpool/taskqueue APIs to peloton/private
Summary:
This change moves offerpool and taskqueue APIs to peloton/private/hostmgr
and peloton/private/resmgr respectively. This allows us to serve those
two APIs in hostmgr and resmgr binaries without significant changes
to the code.

The next step is to define and implement the new APIs for hostmgr and
resmgr. In the meanwhile, we can proceed to execise the job creation
code path using the new architecture with separate binaries.

Test Plan: make test

Reviewers: slu, zhitao, mabansal, #peloton_meta_framework, wu

Reviewed By: #peloton_meta_framework, wu

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D727168