Upgrade packer based vagrant for peloton.
Summary:
1. Upgrade base box from ubuntu/trusty64 to v0rtex/xenial64, which ships latest kernel, overlay filesystem support, and systemd;
2. Update mesos agent flags to support unified containerizer and task group.

I plan to consolidate this as default testing environment for anything related to integrated testing between scheduler, mesos and containerizer.

I will work with Min on upload the box once this is approved.

Test Plan:
Follow the updated instructions in packer build for a new test box, launch it, and verifies that a test run of `mesos-execute` can start a task group of 2 with command:

https://code.uberinternal.com/P38197

Reviewers: min, wu, mu, xiaojian, #peloton, jlai

Reviewed By: jlai

Subscribers: jenkins, rcharles, jlai

Differential Revision: https://code.uberinternal.com/D791905