[API] Return child resource pools in GetResourcePool method
Summary:
This change adds a new includeChildPools param in GetResourcePool
method so that we can return the list of child resource pool infos
if the param is true. This will reduce the extra RPC calls in
Peloton UI when a user clicks through the resource pool hierarchical.

Test Plan: make test

Reviewers: damon, #peloton, mabansal, akhera

Reviewed By: #peloton, mabansal, akhera

Subscribers: mabansal, zhitao, jenkins

Differential Revision: https://code.uberinternal.com/D822161