Symlink peloton logs in container to /var/log/peloton on host
Summary:
To allow Filebeat to ingester Peloton logs, we need to provide a fixed
file path on the host for the console logs in Mesos sandbox. This change
sets up the symlink on the Mesos host at /var/log/peloton/{app_id}
before we launch the Peloton apps.

Test Plan:
(1) deploy Peloton with the change and verify /var/log/peloton will be
created automatically and the corresponding Peloton logs are symlinked
correctly.

Reviewers: wu, zhitao, yunpeng, mu, #peloton

Reviewed By: wu, yunpeng, mu, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D828783