Add generic Queue and Pool to common/ and use in placement/
Summary:
The Pool is very useful for running a large number of jobs on a smaller set of
goroutines, such that the system resources will remain constant when traffic
peak.

The Queue will most likely change significantly going forward, but for now it
offers a nice way of seperating the Queue logic from the Pool.

Reviewers: min, mabansal, #peloton, kejlberg

Reviewed By: kejlberg

Subscribers: jenkins, xiaojian

Differential Revision: https://code.uberinternal.com/D882461