Add null check in job runtime updater.
Summary:
Add null check in job runtime updater.

In devel cluster we are seeing the case that a job runtime record exists in
the DB but the corresponding jobconfig does not exist. Thus the code to update
job runtime would panic.

This change adds the null check as a safe guard. I am investigating the cause of the dangling job_runtime record separately.

Test Plan: Repo the scenario with pcluster and veridfied the fix

Reviewers: wu, #peloton

Reviewed By: wu, #peloton

Subscribers: zhitao, jenkins, min, mu, xiaojian

Differential Revision: https://code.uberinternal.com/D858925