MVP Gang Scheduling Support for Gang Resource Pool Admission Control
Summary:
MVP Gang Scheduling Support for Gang Resource Pool Admission Control

OVERVIEW
This change set adds MVP support for gang scheduling, focusing on gang resource pool admission control.
In this MVP gang scheduling change set, each PENDING queue entry is changed from a task to a scheduling
unit, which is a list of 1 or more tasks that should be admitted as a gang.
BONUS CHANGE: Makefile support for running a subset of the tests; useful during development.

CAVEATS
This change set only changes the PENDING queue and not the READY queue.  It does not ensure that the
tasks forming a gang are inserted and removed from the READY queue as a group, meaning that a gang
that passes admission control but fails to be placed completely (due to fragmentation or constraints)
is not guaranteed to start execution together.  Also, this change set does not enforce gang running-state
semantics, i.e., it does not enforce terminating all running gang members if the entire gang cannot be
kept executing.  Addressing these issues are slated for future change sets; this MVP change set is
intended to allow experimentation with distributed tensorflow job scheduling in the meantime.

DETAILS
The change set contains the following updates in JobManager and ResManager.

Job Manager
+Adds new instance-related sla fields (done in previous checkin but there was a typo)
 protobuf/peloton/api/job/job.proto
+Changes job manager to validate the values of the new instance-related sla fields
 jobmgr/task/config.go
 jobmgr/job/handler_test.go
Note: job mgr already sends all tasks created for a job to ResMgr in one call (ref: jobmgr/job/Create)

Res Manager
+Changes resmgr pending queue entry to specify new minInstance field as > 1 for gang members
 protobuf/peloton/private/resmgr/resmgr.proto
 util/util.go
+Changes priority queue scheduling unit from task to task list
 resmgr/queue/priority.go
 resmgr/queue/priority_test.go
 resmgr/queue/queue.go
+Changes ResMgr service to enqueue gang tasks as single scheduling unit in the PENDING queue
 (although it already received them one call, it was inserting them one at a time up to a limit)
 resmgr/handler.go
 resmgr/handler_test.go
+Changes ResMgr DB load to enqueue gang tasks as single scheduling unit in the PENDING queue
 resmgr/respool/restree.go
 resmgr/respool/restree_test.go
+Changes Scheduler DequeueTasks to dequeue gang tasks as continguous set from PENDING queue
 resmgr/respool/respool.go
 resmgr/respool/respool_test.go
 resmgr/task/scheduler.go

Test Plan:
 make test [DONE]
 integration test with gang scheduling [TBD]

Reviewers: min, #peloton, mabansal

Reviewed By: min, #peloton

Subscribers: zhitao, mu, slu, kejlberg, jenkins, xiaojian

Differential Revision: https://code.uberinternal.com/D910687