Change private interface to resmgr pending queue from list of tasks to list of gangs, each of which is a list of tasks
Summary:
Change private interface to resmgr pending queue from list of tasks to list of gangs, each of which is a list of tasks.
This is a step on the path to updating the task list interface to gang list interface between peloton private services where appropriate.

Address lint problems

Test Plan: make test

Reviewers: min, mabansal, zhitao, mu, #peloton, ajohnsen

Reviewed By: #peloton, ajohnsen

Subscribers: ajohnsen, jenkins, xiaojian

Differential Revision: https://code.uberinternal.com/D950049