Expose filter result in hostmgr interface.
Summary:
It is often asked why hostmgr did not return enough host offers to placement engine, and this is pretty hard to debug.

Given that the `Matcher` object already tracks these reasons, this diff returns that in a map to placement engine so both caller and callee could track these numbers.

Also rename `Constraint` to `Filter`, and add a couple of fields which I plan to implement in the future.

Test Plan: make test

Reviewers: mu, aholler, xiaojian, #peloton

Reviewed By: mu, #peloton

Subscribers: jenkins, xiaojian

Differential Revision: https://code.uberinternal.com/D967719