Initialize the task stats for a job to consider all tasks as initialized
Summary:
Initialize the task stats for a job to consider all tasks as initialized

Before this fix the task stats is initialized to an empty map, and the stats table is updated
by a timer callback later; so there will be a short time period that the task stats is all empty.
This change init the stats map to consider all tasks in "initialized" state.

Reviewers: min, #peloton

Reviewed By: min, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D981963