Implement basic storage operations for workflows
Summary:
For a workflow, we persist two entries to track progress; `Instances` and `Progress`.

- Instances is the set of instance IDs currently being upgraded.
- Progress is the count of instances that has been initiate.

To operate using these principles, an upgrades must process from instance 0 to N, in order.

The state of the upgrade can be explained using these properties:

- An upgrade can start the next task, if `len(Instances) < batch_size`.
- An upgrade has completed `Progress - len(Instances)` tasks.
- An upgrade is done if `Progress == InstanceCount &&  len(Instances) == 0`.
- A task has been upgraded id `InstanceID < Progress && !Instances.Contains(InstanceID)`

The `StartTaskUpgrade` uses row-level transaction to ensure only instance with ID equal to `Progress` can be started, and that any task can only be started once.

Reviewers: min, mu, #peloton, slu

Reviewed By: mu, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1022249