Peloton ELK: fix dynamic keys in logs
Summary:
The log field `"all_offers"` (from the log message `"Offer removed while not in ready status"`) represents `hostSummary.unreservedOffers`, which is of type `map[string]*mesos.Offer`, the key of the map being the OfferID:
```
"all_offers": {
              "6b4cc7c2-7249-4df5-a94e-854db20958f5-O2823816": {
                "agent_id": {
                  "value": "7d8dd8b0-7ffa-4418-84b8-339f3c1c0ca5-S146"
                },
                "allocation_info": {
                  "role": "peloton"
                },
...
```

With this behavior, every new offerID is a new key in the log message (from ELK perspective) and will be categorized by ELK as a `dynamic key`, which will burst the `maximim limit of 1000 keys per index`. As a result if a new key is introduced (which is the case for https://code.int.uberatc.com/D58678) then no document will get indexed, hence the dashboard will be empty of documents...

A solution is to blacklist from ELK indexing this `"all_offers"` field by using the `_noindex suffix` and so to rename the field `"all_offers" -> "all_offers_noindex"`
(https://engdocs.uberinternal.com/ELKLogSearch/produce_logs.html#what-gets-indexed).

Reviewers: #peloton, min, mehrotra, zhitao, mu

Reviewed By: #peloton, zhitao, mu

Subscribers: jenkins, surajs

Differential Revision: https://code.uberinternal.com/D1108655