Untrack tasks when they become unactive
Summary:
This includes a minor cleanup for how locks are used. The "job lock" will eventually move into the tracked manager, such that we split the responsibility of

1) Suggesting an action
2) Decide if the action should be taken

All changes to the manager is thus taken while holding the manager lock, while the job/task locks ensures a very fast read-only view to the jobs/tasks.

Reviewers: fgalilee, kejlberg, #peloton

Reviewed By: fgalilee

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1150405