Don't lock runtime updater while processing job
Summary:
When a job runtime is being updated for task stats, a global lock is taken that's also shared with mesos event processing. The result is that mesos events won't be processed while any job is being processed. When benchmarking jobs with 100k tasks, the interval for updating job runtimes is quite significant. That, combine with the change in D1131907, would mean we had a global block on mesos events.

Instead of relaying on the lock to ensure the job is only proceessed once, we use a map of channels to communicate a single processing and blocking.

This is a potential fix for T1141691.

Reviewers: min, mu, wu, fgalilee, #peloton, slu, mabansal

Reviewed By: fgalilee

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1155699