M3 config for cluster tag is not being correctly overriden by CLUSTER environment variable
Summary: The M3 configuration should override the common tag "cluster" if the "CLUSTER" environment varible is set. This tag is set to the default value of "peloton-default". However, the loadM3Configs() function is using using the environment variable name (which is upper-case) as the key in the CommonTags map. This creates two (k,v) entries: "(cluster, peoloton-default)" and "(CLUSTER, real-cluster-name)". When metrics are emitted by M3, they are emitted twice with the two values for "cluster".

Reviewers: jpinkul, mabansal, #peloton, leonr

Reviewed By: jpinkul, mabansal, #peloton, leonr

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1157771