Refactor the placement mappers and models
Summary: Moves all the code in the `placement.mappers` package into `placement.models` and adds methods on the models for the `Offer` and `Task` types to lazily convert them to the equivalent Mimir types `Group` and `Entity`, respectively. This will be needed by a later diff to refactor the `placement.offers.Manager` and the `placement.tasks.Manager` which will again later be used to refactor the main placement engine loop to de-couple the logic to fetch tasks and offers, set placements and return unplaced tasks and unused offers and the logic that decides which offers to assign to which tasks.

Reviewers: fgalilee, ajohnsen, #peloton, mu

Reviewed By: ajohnsen, #peloton

Subscribers: zhitao, min, jenkins

Maniphest Tasks: T1021963

Differential Revision: https://code.uberinternal.com/D1156775