Ensuring state count metrics do not go negative, and that we do not update counts for invalid transisions
Summary:
In WBU2, we were seeing task counts metrics that were negative for some states. This is most likely due to the fact that the update logic
for the task counts does not check if the "from" task count is > 0. There is also the case where invalid transisions still update the task
counts.

{F25988235}

Reviewers: jpinkul, leonr, min, mabansal, #peloton

Reviewed By: jpinkul, leonr, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1169781