Add revision(version) field to task runtime
Summary:
With this change we return an (yarpc) error indicating a conflict, such
that a retry can be attempted.

A retry should start from scratch; by reading the latest version from
DB, applying the change and try store it again (if still applicable).

This diff was already landed on the `goalstate` branch. Getting this landed
on master will effective close the `goalstate` branch.

Reviewers: min, mu, slu, fgalilee, kejlberg, schwarz, #peloton

Reviewed By: kejlberg, #peloton

Subscribers: mabansal, eskil, jenkins

Differential Revision: https://code.uberinternal.com/D1169711