Add a integration test that runs a 100 tasks with different resource constraints
Summary: This diff adds a single integration test that tries to run a batch job with a 100 tasks that all have different task configs. The purpose of this is to test how the placement works in the presence of many different tasks with different resource constraints that should be place concurrently.

Reviewers: ajohnsen, fgalilee, #peloton

Reviewed By: ajohnsen, #peloton

Subscribers: jenkins

Maniphest Tasks: T1021963

Differential Revision: https://code.uberinternal.com/D1185909