[API] Refactor job, task and respool service APIs
Summary:
This change refactors the job, task, and respool service APIs
so that we are switching to use standard gRPC error codes.
https://github.com/grpc/grpc/blob/master/doc/statuscodes.md

This will allow use to report API error code metrics via YARPC.

Also moved the service definition into a separate package so that
we can keep the common message types in job.proto, task.proto
and respool.proto. The package naming convention we are going to
use is the following:

```
peloton/
   api/
      job/
         job.proto          <== common message types
         svc/
            job_svc.proto   <== service definitions w/ request
                                and response messages.
```

Test Plan: (1) unit tests and integration tests

Reviewers: zhitao, #peloton, ajohnsen, mu

Reviewed By: #peloton, ajohnsen, mu

Subscribers: mabansal, leonr, avyas, jenkins

Differential Revision: https://code.uberinternal.com/D1123285