Removing tasks from the pending queue during admission control if they are killed.
Summary:
If the job/tasks are killed they are removed from the tracker however if the tasks
in the pending queue in resmgr and cluster does not have the resources which task wants
those tasks can block the whole queue. This diff addresses this issue while doing the
admission control it checks the validity of the tasks and if the task is not valid then
it removes the task from the queue. We do this in admission control for not taking the big
lock while killing, otherwise that will have a big impact on the throughput while killing at
the same time

Test Plan: Test added, Tested with pcluster

Reviewers: avyas, #peloton

Reviewed By: avyas, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1210047