Adds task preemptor for RUNNING tasks in jobmgr
Summary:
Task preemption for RUNNING tasks is the responsibility
of the job manager. The preemptor module in jobmgr will periodically
get the list of tasks to preempt from the resource manager. It will
then process each task so that they are `KILLED` and the rescheduled
into the resource manager with a new mesos task ID. This diff
includes changes to
1. Add a task preemptor goroutine which will gets the tasks to preempt from the resource manager
2. Goal state engine changes to act on the `PREEMPTING` goal state
3. Add a new `Initialize` action to be used by the goalstate engine
4. Tests

Reviewers: #peloton, mabansal, ajohnsen

Reviewed By: #peloton, mabansal, ajohnsen

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1222397