Merge the placement engine from the master branch into release.
Reviewers: fgalilee, #peloton

Differential Revision: https://code.uberinternal.com/D1371127

Add missing changes from Makefile from master.

Merge changes to the daemon.

Review comments.

Fix test and rename tests and constructors to reflect the rename of the type models.Offer to models.Host.

Merge the coordinator config into the placement config.

Rename offer to host in the assignment.

Merge the coordinator into the engine.

Review comments: logging and renaming.

Make task max rounds and max duration for placement configurable.

Add extra tests and make the strategy decide if it is concurrency safe or not.

Rebase and use the task type flag.

Review comments.