Process event listeners after primary handling
Summary:
This fixes a race where we would update the job runtime before we had applied the task event update:

1) mesos event is received
2) task event is processed and task status is updated
3) job runtime update is triggered and handled (task state in job is updated)

As the events was processed in parallel, there was no ordering between 2 and 3, and could thus happen in reversed order. The result is we occasionally not have a consistent view within the job state / drop an event.

Reviewers: kejlberg, #peloton, schwarz

Reviewed By: kejlberg, #peloton, schwarz

Subscribers: jenkins

Maniphest Tasks: T1189709

Differential Revision: https://code.uberinternal.com/D1205173