Handle Cassandra errors in the storage layer
Summary:
This is part one of a multi-part commit.
This commit adds code in the storage layer to handle Cassandra errors.
Errors are classified into the following categories.

1. Retry in storage.
Try 5 times in the storage layer with a backoff, and if retry fails
all 5 times, push it up to the caller with the error being
a TransientError. Connection error is an example of such errors.
2. Retry in caller.
Push the error to the caller as a TransientError.
WriteTimeout error is an example of such errors.
3. Dont do anything.
Just push the error as is to the caller.
Syntax errrors are example of such errors.

Code to handle TransientErrors in the caller still needs to be added.

Reviewers: mu, mabansal, #peloton, ajohnsen

Reviewed By: mu, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1276521