Add additional logs when AlreadyExistsErrorf occurs
Summary:
In https://code.uberinternal.com/T1268797, the task update
was stuck in a seemingly infinite loop with error AlreadyExistsErrorf.
To debug future such occurences, adding logs in jobmgr with
respect to what is being added to the DB.
What is present in the DB can be derived from looking in
the Cassandra DB, and hence we can figure out the exact
operation which is failing.

Reviewers: mabansal, ajohnsen, #peloton

Reviewed By: ajohnsen, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1283129