Use non synchronized `isLeaf` in `respool.CalculateAllocation`
Summary:
Any numbers of goroutines may simultaneously acquire a
RLock of the same RWMutex. What does not work is when the same
goroutine calls RLock on a RWMutex when that goroutine already
did that without calling RUnlock first.

https://go-review.googlesource.com/#/c/23570/

This could be the possible cause of the deadlock we see in production.
In this particular case `ToResourcePoolInfo` acquires a read lock and
calls `IsLeaf` which again tries to acquire a read lock on the same mutex.
This diff changes the call to the non sychronized version `isLeaf`. In
any case we should fix this.

Reviewers: mabansal, min, #peloton

Reviewed By: mabansal, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1297453