Add per task preemption policy
Summary:
When a running task is preempted, add an option
to specify if the task should be rescheduled or not. This diff
adds a new `PreemptionPolicy` inside `TaskConfig` inside which
one can specify `killOnPreempt`. If `killOnPreempt` is set to
`true` , once the task is preempted it won't be rescheduled, it
is up to the client to handle preemption. The default is set to
false where after preemption the task will be rescheduled.

Test Plan: Added tests

Reviewers: #peloton, apoorvaj, mabansal

Reviewed By: #peloton, apoorvaj

Subscribers: apoorvaj, jenkins

Differential Revision: https://code.uberinternal.com/D1338479