Merge the placement engine from the dev branch into master.
Summary:
This diff merges the placement engine refactoring from the `dev` branch into the `master` branch. This diff contains the following changes:
   - bugfixes to async/daemon.go
   - placement/models changes
   - task and offer service changes
   - placement engine refactoring to accommodate new placement strategies.
   - batch plugin (the current batch placement logic)
   - mimir plugin (the new Mimir placement logic, which will start by doing stateful placement).

Performance testing:
I rebased this on the release-0.4.0 branch and ran the load tests on the `devel01` cluster. For `HEAD` I got these results:
  - 259.51 seconds
  - 255.38 seconds
  - 250.04 seconds
  - 245.09 seconds
  - 242.93 seconds
Which give an average of  250.59 seconds.

For this diff I got these results:
  - 239.89 seconds
  - 247.49 seconds
  - 245.31 seconds
  - 249.78 seconds
  - 262.10 seconds
Which give an average of  248.91 seconds.

Reviewers: fgalilee, #peloton, mu

Reviewed By: #peloton, mu

Subscribers: zhitao, min, jenkins

Differential Revision: https://code.uberinternal.com/D1260881