Resource Manager recovery should not recover Initialized jobs and not fail if some tasks cant be recovered
Summary:
Right now Resourcemanager recovers Initalized job states which it should not do as jobmanager already does that.
In recovery right now resource manager fails if it fails to enqueue any task which ot should not and should move forward.
Right now we are logging error in this case and sending it to sentry also updating new metrics , which we ill use to
alert

Test Plan: modified

Reviewers: avyas, #peloton

Reviewed By: avyas, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1343861