Handle stopped tasks during reconcialiation.
Summary:
If a LOST message is received for a task with goal state KILLED,
just mark the task state as KILLED. The same message is also sent to
resource manager which will clear the placement.
And mesos has already cleaned up the task. So there is nothing
else which job manager needs to do.

Resolves T1394395 T1396937

Reviewers: mu, zhitao, #peloton

Reviewed By: mu, #peloton

Subscribers: jenkins, zhitao

Maniphest Tasks: T1396937, T1394395

Differential Revision: https://code.uberinternal.com/D1362761