Store default config in job config and instance config in task config
Summary:
Also, fetch the configuration only when the caller really needs to use it.
This reduces the amount of read and write data to Cassandra per query.
Without this change, for a job configuration with multiple instances, and each
instance having its own configuration, the job and and instance configs being
stored in the DB can become unnecessarily huge.

Introduce job goal state structure in proto for future use.
It is currently used to delete jobs for which jobv was created
successfully, but the task configs failed to create.

Resolves T1410133 T1286239

Reviewers: min, mabansal, mu, #peloton

Reviewed By: mu, #peloton

Subscribers: avyas, zhitao, jenkins

Maniphest Tasks: T1410133, T1286239

Differential Revision: https://code.uberinternal.com/D1318307