Return failed instance list in task stop and task start
Summary:
In case DB update fails for a task, add it to the failed instance
list and return the list in the response.
Also retry once in case of task stop in case a DB update
failed for a few tasks.
Also add a "job stop" in cli which stops all tasks in a job.

Resolves T1285221 T1183637 T1316995

Reviewers: mu, min, #peloton

Reviewed By: mu, #peloton

Subscribers: jenkins

Maniphest Tasks: T1316995, T1285221, T1183637

Differential Revision: https://code.uberinternal.com/D1367439