Update volume state upon stateful task running status update
Summary:
We always batch RESERVE->CREATE->LAUNCH offer operations together when
we launch a stateful task. volume will be marked as CREATED if this task is running.
Previous we were doing that in hostmgr, which is too late since the volume will be marked
as CREATED only after the task is stopped.

Reviewers: apoorvaj, #peloton, zhitao

Reviewed By: #peloton, zhitao

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1392705