Implement job recovery in goal state engine in job manager
Summary:
This commit contains the following pieces of code.
1. Abstract the fast job recovery code in resource manager to a common infra.
2. Have resource manager use this common infra.
3. Implement recovery in goal state in job manager using this common infra.
All non-terminal jobs need to be recovered.
In addition, failed or killed jobs with non-terminal goal states
also needs to be recovered.
All recovered jobs and tasks are loaded into goal state for evaluation.
4. Start host manager event stream only after goal state recovery is complete.
This ensures that no mesos events are processed before goal state is in sync
with the DB. This eliminates any potential race conditions which can happen
if mesos events update the goal state while it is syncing with the DB.
5. Race conditions can still exist between updates from API during DB sync.
Starting API handler after goal state db sync will lead do a downtime during
upgrade, hence not using that option.
Instead, the DB sync checks if the job or the task already exists, then keep
the old state instead of loading the one from DB. This should eliminate all races,
but needs to be tested in staging and production carefully.

Resolves T1435191

Test Plan:
1. Unit and integrations tests pass.
2. Tested the recovery locallhy using pcluster.

Reviewers: min, mabansal, mu, avyas, #peloton

Reviewed By: avyas, #peloton

Subscribers: jenkins

Maniphest Tasks: T1435191

Differential Revision: https://code.uberinternal.com/D1399727