Add support for MaximumRunningInstances SLA configuration.
Summary:
MaximumRunningInstances limits the maximum number of concurrent
instances which can be running in a job.
This SLA can be broken by a customer action of starting
terminated tasks using task Start API.
The code brings in the following pieces to support the feature.
1. Limit scheduling of tasks when job first created.
2. Maintain the SLA during job recovery.
3. Maintain the SLA after job manager restart.
4. When a task reaches its goal state, it is untracked from the
manager. Re-evaluate the job SLA to schedule more instances
if a task gets untracked in goal state.

Resolves T952969

Test Plan:
1. Unit and integration tests pass.
2. Tested for different values of maximumrunninginstances in local pcluster.
3. More testing to be done in preprod after landing the change.

Reviewers: avyas, mu, min, mabansal, #peloton

Reviewed By: avyas, #peloton

Subscribers: jenkins

Maniphest Tasks: T952969

Differential Revision: https://code.uberinternal.com/D1410911