Fix sorting based on creation/completion time in job query
Summary:
The commit contains the following.

Add time in addition to the date in the lucene index for all variables of type time
Add a new index with this change. In the next migration, the original index will be dropped.

Return the original query spec in the query response.

Do not store the instance configuration with the job configuration
in the job index table. Hence, the instance configuration is no longer queriable.

If no sorting parameter is provided, sort by default on creation_time in descending order.

Resolves T1477397 T1462697

Reviewers: mu, mabansal, #peloton, min

Reviewed By: #peloton, min

Subscribers: derek.leslie, jenkins

Maniphest Tasks: T1477397, T1462697

Differential Revision: https://code.uberinternal.com/D1436403