A task in launched state times out in job manager
Summary:
Currently, after a task is dequeued by job manager, a timer for 20 minutes
starts in resource manager. This timer times out both job manager delays as well as mesos taking too much time to start a task.
This behavior is being changed as follows.
After job manager launches the task, it sends a notification to resource manager which will transition the task state to
LAUNCHED and cancel the timer.
Then, job manager starts a 20 minute timeout in case mesos fails to launch in 20 minutes.
If the launch times out, the mesos id is regenerated and the task is sent again to resource manager for placement.

The reason to make this change is because after a task has been launched, it should not be launched
again with the same mesos id otherwise the second launch will be ignored by mesos.
Since job manager does the launch, and it is the one which updates the DB with new mesos id,
it needs to be one to timeout after launch.

Resolves T1476915 T1476939 T934391

Reviewers: mu, #peloton, avyas

Reviewed By: #peloton, avyas

Subscribers: min, mabansal, jenkins

Maniphest Tasks: T1476939, T1476915, T934391

Differential Revision: https://code.uberinternal.com/D1470023