Use the version in job configuration protobuf to track job version
Summary:
Currently, the version field in the job_config table in DB is the
only location to hold the job configuration version, and it is
stored in job runtime protobug as well. None of other places
including task runtime, task config and job config protobufs
store and use this value.
With this commit, the job version will be tracked in the job
configuration protobuf. It will be copied to job runtime
protobuf, task runtime protobuf, job config table in DB
and task config table in DB.

Resolves T1774561

Test Plan: Tested create and update jobs using pcluster.

Reviewers: mu, min, #peloton, zhixin

Reviewed By: #peloton, zhixin

Subscribers: mabansal, jenkins

Maniphest Tasks: T1774561

Differential Revision: https://code.uberinternal.com/D1752251