Adding peloton engdocs
Summary:
This diff is adding the first version of Eng Docs in the repo.

Eng Docs Creation
===============
For Eng docs to be created we need to do the following steps

1. Tag the repo as foober
2. Create the docs folder inside your code repo
3. Add the doc template code.
4. Change the whole template based on your project documentation
5. Build the doc locally "foober docs make"
6. Verify the Doc
7. arc diff and arc land the diff

Once the diff is landed, Jenkins job will run and build the docs and publish the docs to eng docs.

Update EngDocs
=============
Once we have first version of docs in the repo, Follow below steps to update the docs

1. checkout the docs folders and make the change in the doc
2. arc diff and arc land
3. Eng docs will be updated

Reviewers: avyas, min, #peloton

Reviewed By: avyas, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1474421