Schema and DB change to speed up JobQuery
Summary:
Job Query takes a lot of time to complete for
clusters which have jobs with huge job config.
Initial suspicion was lucene index being slow, but that is not
the case. I executed the lucene index query to get job id's
manually on pit1-prod01 for /Simulations RP (for which
the Job Query request for completed jobs times out)
That query completes in under 5sec. The main problem is that
for each of the 100 job id's returned by lucene index,
we query the job_index table for their respective job config,
which can be pretty huge for each job. This slows down the
over all query.

It turns out, that for job query, we don't really need the
entire config, if we add job name and runtime info to the
lucene index.

This change sets the stage for deprecating returning config
in job query response as well as makes the schema change
for lucene index.

Reviewers: #peloton, min, apoorvaj, mu

Reviewed By: #peloton, min, apoorvaj

Subscribers: rcharles, jenkins

Differential Revision: https://code.uberinternal.com/D1474883